ADTS-272 Add type files
