+++ src/Components/OATPropertyEditor/AddPropertyBar.types.ts
@@ -1,0 +1,5 @@
+export type AddPropertyBarProps = {
+    onMouseOver?: (event: React.MouseEvent<Element, MouseEvent>) => void;
+    onClick?: () => void;
+    classNameIcon?: string;
+};
