+++ src/Components/OATPropertyEditor/Editor.types.ts
@@ -1,0 +1,14 @@
+import { IAction } from '../../Models/Constants/Interfaces';
+import { IOATPropertyEditorState } from './OATPropertyEditor.types';
+import { IOATEditorState } from '../../Pages/OATEditorPage/OATEditorPage.types';
+import { IDropdownOption } from '@fluentui/react';
+import { Theme } from '../../Models/Constants/Enums';
+
+export type EditorProps = {
+    dispatch?: React.Dispatch<React.SetStateAction<IAction>>;
+    dispatchPE?: React.Dispatch<React.SetStateAction<IAction>>;
+    languages: IDropdownOption[];
+    state?: IOATEditorState;
+    statePE?: IOATPropertyEditorState;
+    theme?: Theme;
+};
