+++ src/Components/OATPropertyEditor/PropertyListItemSubMenu.tsx
@@ -5,25 +5,7 @@ import {
     getPropertyInspectorStyles,
     getIconMoreSubMenuItemStyles
 } from './OATPropertyEditor.styles';
-
-type IPropertyListItem = {
-    index?: number;
-    subMenuActive?: boolean;
-    deleteItem?: (index: number) => void;
-    deleteNestedItem?: (parentIndex: number, index: number) => void;
-    parentIndex?: number;
-    onDuplicate?: () => void;
-    onPropertyListAddition?: () => void;
-    onTemplateAddition?: () => void;
-    onMoveUp?: (index: number, moveUp: boolean) => void;
-    onMoveDown?: (index: number, moveUp: boolean) => void;
-    removeItem?: boolean;
-    duplicateItem?: boolean;
-    addItemToPropertyList?: boolean;
-    addItemToTemplates?: boolean;
-    targetId?: string;
-    setSubMenuActive: React.Dispatch<React.SetStateAction<boolean>>;
-};
+import { PropertyListItemSubMenuProps } from './PropertyListItemSubMenu.types';
 
 export const PropertyListItemSubMenu = ({
     index,
@@ -42,7 +24,7 @@ export const PropertyListItemSubMenu = ({
     addItemToTemplates,
     targetId,
     setSubMenuActive
-}: IPropertyListItem) => {
+}: PropertyListItemSubMenuProps) => {
     const { t } = useTranslation();
     const propertyInspectorStyles = getPropertyInspectorStyles();
     const subMenuItemStyles = getIconMoreSubMenuItemStyles();
