+++ src/Components/OATPropertyEditor/PropertyListMapItemNested.types.ts
@@ -1,0 +1,9 @@
+import { IAction } from '../../Models/Constants/Interfaces';
+import { IOATEditorState } from '../../Pages/OATEditorPage/OATEditorPage.types';
+
+export type PropertyListMapItemNestedProps = {
+    dispatch?: React.Dispatch<React.SetStateAction<IAction>>;
+    index?: number;
+    item?: any;
+    state?: IOATEditorState;
+};
