+++ src/Components/OATPropertyEditor/TemplateList.tsx
@@ -9,20 +9,9 @@ import {
     SET_OAT_CONFIRM_DELETE_OPEN,
     SET_OAT_PROPERTY_EDITOR_DRAGGING_TEMPLATE
 } from '../../Models/Constants/ActionTypes';
-import { IAction } from '../../Models/Constants/Interfaces';
-import { IOATEditorState } from '../../Pages/OATEditorPage/OATEditorPage.types';
-import { getModelPropertyCollectionName } from './Utils';
 
-interface ITemplateList {
-    draggingTemplate?: boolean;
-    draggingProperty?: boolean;
-    enteredTemplateRef: any;
-    draggedTemplateItemRef: any;
-    enteredPropertyRef: any;
-    dispatch?: React.Dispatch<React.SetStateAction<IAction>>;
-    state?: IOATEditorState;
-    dispatchPE?: React.Dispatch<React.SetStateAction<IAction>>;
-}
+import { getModelPropertyCollectionName } from './Utils';
+import { TemplateListProps } from './TemplateList.types';
 
 export const TemplateList = ({
     draggedTemplateItemRef,
@@ -33,7 +22,7 @@ export const TemplateList = ({
     dispatch,
     dispatchPE,
     state
-}: ITemplateList) => {
+}: TemplateListProps) => {
     const { execute } = useContext(CommandHistoryContext);
     const propertyInspectorStyles = getPropertyInspectorStyles();
     const dragItem = useRef(null);
