+++ src/Pages/OATEditorPage/Internal/Components/OATModal.tsx
@@ -1,13 +1,8 @@
 import React from 'react';
 import { Modal as FluentModal } from '@fluentui/react';
+import { OATModalProps } from './OATModal.types';
 
-interface IModalProps {
-    children?: React.ReactNode;
-    isOpen?: boolean;
-    className?: string;
-}
-
-export const OATModal = ({ children, isOpen, className }: IModalProps) => {
+export const OATModal = ({ children, isOpen, className }: OATModalProps) => {
     return (
         <FluentModal isOpen={isOpen} containerClassName={className}>
             {children}
