+++ src/Pages/OATEditorPage/Internal/Components/OATModal.types.ts
@@ -1,0 +1,5 @@
+export type OATModalProps = {
+    children?: React.ReactNode;
+    isOpen?: boolean;
+    className?: string;
+};
